Create Components and Set up Router in App.js
Create HedgeSimpleML.js Takes in user input and gives a suggested hedge bet to have equal payouts. The function for this was created in the algos folder and is called simpleML.js

Create HedgeSimpleSpread.js Similar to HedgeSimpleML.js takes in user input and the algo for that is in the spreadHedge.js file

Create Suggestions.js this is a component for the HedgeSimpleML.js to pass its information to and render a suggestion of what to bet.

Create Navbar.js to be able to switch between ML hedges and Spread heades.

App.js imported the Router so that the Navbar could be used.